third_party/WebKit/LayoutTests/inspector/console/console-functions-expected.txt

CONSOLE MESSAGE: line 24: function simple() {}
CONSOLE MESSAGE: line 25: function simple() {}
CONSOLE MESSAGE: line 24: async function asyncSimple() {}
CONSOLE MESSAGE: line 25: async function asyncSimple() {}
CONSOLE MESSAGE: line 24: function* genSimple() {}
CONSOLE MESSAGE: line 25: function* genSimple() {}
CONSOLE MESSAGE: line 24: function (){}
CONSOLE MESSAGE: line 25: function (){}
CONSOLE MESSAGE: line 24: function (x, y){}
CONSOLE MESSAGE: line 25: function (x, y){}
CONSOLE MESSAGE: line 24: function namedArgs(x) {}
CONSOLE MESSAGE: line 25: function namedArgs(x) {}
CONSOLE MESSAGE: line 24: function namedArgs2(x, y) {}
CONSOLE MESSAGE: line 25: function namedArgs2(x, y) {}
CONSOLE MESSAGE: line 24: function ({}) {}
CONSOLE MESSAGE: line 25: function ({}) {}
CONSOLE MESSAGE: line 24: function* whitespace(  x  )    {   }
CONSOLE MESSAGE: line 25: function* whitespace(  x  )    {   }
CONSOLE MESSAGE: line 24: async function whitespace2(  x  ,  y  ,  z  )    {   }
CONSOLE MESSAGE: line 25: async function whitespace2(  x  ,  y  ,  z  )    {   }
CONSOLE MESSAGE: line 28: [object Object]
CONSOLE MESSAGE: line 29: [object Object]
Tests that console logging different types of functions correctly.

console-functions.html:24 ƒ simple() {}
console-functions.html:25 ƒ simple()
    arguments: null
    caller: null
    length: 0
    name: "simple"
    prototype: {constructor: ƒ}
    __proto__: ƒ ()
    [[FunctionLocation]]: console-functions.html:10
    [[Scopes]]: Scopes[1]
console-functions.html:24 async ƒ asyncSimple() {}
console-functions.html:25 async ƒ asyncSimple()
    arguments: (...)
    caller: (...)
    length: 0
    name: "asyncSimple"
    __proto__: AsyncFunction
    [[FunctionLocation]]: console-functions.html:11
    [[Scopes]]: Scopes[1]
console-functions.html:24 ƒ* genSimple() {}
console-functions.html:25 ƒ* genSimple()
    arguments: (...)
    caller: (...)
    length: 0
    name: "genSimple"
    prototype: Generator {}
    __proto__: GeneratorFunction
    [[FunctionLocation]]: console-functions.html:12
    [[IsGenerator]]: true
    [[Scopes]]: Scopes[1]
console-functions.html:24 ƒ (){}
console-functions.html:25 ƒ anonymous()
    arguments: null
    caller: null
    length: 0
    name: ""
    prototype: {constructor: ƒ}
    __proto__: ƒ ()
    [[FunctionLocation]]: console-functions.html:13
    [[Scopes]]: Scopes[1]
console-functions.html:24 ƒ (x, y){}
console-functions.html:25 ƒ anonymous(x, y)
    arguments: null
    caller: null
    length: 2
    name: ""
    prototype: {constructor: ƒ}
    __proto__: ƒ ()
    [[FunctionLocation]]: console-functions.html:14
    [[Scopes]]: Scopes[1]
console-functions.html:24 ƒ namedArgs(x) {}
console-functions.html:25 ƒ namedArgs(x)
    arguments: null
    caller: null
    length: 1
    name: "namedArgs"
    prototype: {constructor: ƒ}
    __proto__: ƒ ()
    [[FunctionLocation]]: console-functions.html:15
    [[Scopes]]: Scopes[1]
console-functions.html:24 ƒ namedArgs2(x, y) {}
console-functions.html:25 ƒ namedArgs2(x, y)
    arguments: null
    caller: null
    length: 2
    name: "namedArgs2"
    prototype: {constructor: ƒ}
    __proto__: ƒ ()
    [[FunctionLocation]]: console-functions.html:16
    [[Scopes]]: Scopes[1]
console-functions.html:24 ƒ ({}) {}
console-functions.html:25 ƒ anonymous({})
    arguments: null
    caller: null
    length: 1
    name: ""
    prototype: {constructor: ƒ}
    __proto__: ƒ ()
    [[FunctionLocation]]: console-functions.html:17
    [[Scopes]]: Scopes[1]
console-functions.html:24 ƒ* whitespace(  x  )    {   }
console-functions.html:25 ƒ* whitespace(  x  )
    arguments: (...)
    caller: (...)
    length: 1
    name: "whitespace"
    prototype: Generator {}
    __proto__: GeneratorFunction
    [[FunctionLocation]]: console-functions.html:18
    [[IsGenerator]]: true
    [[Scopes]]: Scopes[1]
console-functions.html:24 async ƒ whitespace2(  x  ,  y  ,  z  )    {   }
console-functions.html:25 async ƒ whitespace2(  x  ,  y  ,  z  )
    arguments: (...)
    caller: (...)
    length: 3
    name: "whitespace2"
    __proto__: AsyncFunction
    [[FunctionLocation]]: console-functions.html:19
    [[Scopes]]: Scopes[1]
console-functions.html:28 Object
    func0: ƒ simple()
    func1: async ƒ asyncSimple()
    func2: ƒ* genSimple()
    func3: ƒ ()
    func4: ƒ (x, y)
    func5: ƒ namedArgs(x)
    func6: ƒ namedArgs2(x, y)
    func7: ƒ ({})
    func8: ƒ* whitespace(  x  )
    func9: async ƒ whitespace2(  x  ,  y  ,  z  )
    __proto__: Object
console-functions.html:29 Object
    func0: ƒ simple()
    func1: async ƒ asyncSimple()
    func2: ƒ* genSimple()
    func3: ƒ ()
    func4: ƒ (x, y)
    func5: ƒ namedArgs(x)
    func6: ƒ namedArgs2(x, y)
    func7: ƒ ({})
    func8: ƒ* whitespace(  x  )
    func9: async ƒ whitespace2(  x  ,  y  ,  z  )
    __proto__: Object
